requirements.txt
anthropic==0.34.1
beautifulsoup4==4.12.3
cohere==5.8.1
httpx==0.27.0
langchain==0.2.14
langchain_anthropic==0.1.23
langchain_cohere==0.2.2
langchain_community==0.2.12
langchain_core==0.2.33
langchain_openai==0.1.22
openai==1.41.1
pdfminer.six==20231228
pydantic==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
streamlit==1.37.1
toml==0.10.2
duckduckgo-search
